added postsList
